Munge Updates
I created a function that performs the min-max normalization, dummy
transformation and log sale price.  As well, the function has a hard
coded list of columns to be made numeric, which we can adjust.  This
leaves us with quite a bit of dummy attributes, over 2K so there is
definatly some trimming of the data needed.